chore: add skip SSL certificate verification
